fix first bugs with shift masters

+++ app/Http/Controllers/ShiftController.php
@@ -302,7 +302,9 @@ class ShiftController extends Controller
                             continue;
                         }
                     }
-                    $allShift->users()->attach($user->id, ['shift_count' => $collideCount + 1]);
+                    if($allShift->getEmptyUserCountAttribute() > 0){
+                        $allShift->users()->attach($user->id, ['shift_count' => $collideCount + 1]);
+                    }
                 }
             }
         }
@@ -346,7 +348,9 @@ class ShiftController extends Controller
                             continue;
                         }
                     }
-                    $allShift->users()->attach($user->id, ['is_master' => true]);
+                    if($allShift->getEmptyMasterCountAttribute() > 0){
+                        $allShift->users()->attach($user->id, ['is_master' => true]);
+                    }
                 }
             }
         }
@@ -396,7 +400,9 @@ class ShiftController extends Controller
                     }
                 }
                 if ($allShift->id !== $shift->id) {
-                    $allShift->freelancer()->attach($freelancer->id, ['shift_count' => $collideCount + 1]);
+                    if($allShift->getEmptyUserCountAttribute() > 0){
+                        $allShift->freelancer()->attach($freelancer->id, ['shift_count' => $collideCount + 1]);
+                    }
                 }
             }
         }
@@ -433,7 +439,9 @@ class ShiftController extends Controller
                             continue;
                         }
                     }
-                    $allShift->freelancer()->attach($freelancer->id, ['is_master' => true]);
+                    if($allShift->getEmptyMasterCountAttribute() > 0) {
+                        $allShift->freelancer()->attach($freelancer->id, ['is_master' => true]);
+                    }
                 }
             }
         }
@@ -470,7 +478,9 @@ class ShiftController extends Controller
                             continue;
                         }
                     }
-                    $allShift->service_provider()->attach($serviceProvider->id, ['is_master' => true]);
+                    if($allShift->getEmptyMasterCountAttribute() > 0){
+                        $allShift->service_provider()->attach($serviceProvider->id, ['is_master' => true]);
+                    }
                 }
             }
         }
@@ -496,14 +506,16 @@ class ShiftController extends Controller
         $collideCount = $collidingShiftIds->count();
 
 
-        if ($request->chooseData['onlyThisDay'] === false) {
-            $allShifts = Shift::where('shift_uuid', $shift->shift_uuid)
-                ->get();
+        if($request->chooseData !== null){
+            if ($request->chooseData['onlyThisDay'] === false) {
+                $allShifts = Shift::where('shift_uuid', $shift->shift_uuid)
+                    ->get();
 
 
-            foreach ($allShifts as $allShift) {
-                if($allShift->id !== $shift->id){
-                    $allShift->users()->detach($user->id);
+                foreach ($allShifts as $allShift) {
+                    if($allShift->id !== $shift->id){
+                        $allShift->users()->detach($user->id);
+                    }
                 }
             }
         }
